Fixed browser.cache permission and desktop item
Fixed the permission of the installation directory
so that it is games group writeable. This is needed for
the list of ROMs to load correctly. TODO: Instead of
making the directory writeable lets add an initial
browser.cache to files/ and install it to /var/cache and
then symlink it.

Also fixed the desktop entry so that instead of starting
wine directly it starts the shell script wrapper. This will
make it possible to have an apparmor profile for jnes and a
more restrictive one for wine itself.