Added ksshaskpass patch to firefox
This is a work in progress. This patch makes Firefox
use ksshaskpass to prompt for the master password for
the password store. This allows you to save the master
password on kwallet.

At the present the external password tool is hardcoded to
/usr/bin/ksshaskpass but the plan is to make this a configurable
option. If ksshaskpass is not found or cannot be executed for
whatever reason it falls back on the native prompt.

The most important limitation is that if the master password
is stored on the wallet the 'Show Passwords' button will show the
passwords without prompting. This will be addressed in a future
version.